Refactor common package
