+++ common/configuration/BUILD
@@ -1,6 +1,6 @@
 java_library(
     name = "benchmark-configuration",
-    srcs = glob(["*.java"]),
+    srcs = glob(["*.java", "parse/*.java"]),
     deps = [
         "//dependencies/maven/artifacts/com/fasterxml/jackson/core:jackson-core",
         "//dependencies/maven/artifacts/com/fasterxml/jackson/core:jackson-databind",
