+++ common/configuration/parse/BenchmarkArguments.java
@@ -16,8 +16,9 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.benchmark.common.configuration;
+package grakn.benchmark.common.configuration.parse;
 
+import grakn.benchmark.common.configuration.ConfigurationException;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.DefaultParser;
