+++ common/configuration/scenario/complex/queries_complex_write.yml
