+++ common/configuration/scenario/complex/schema.gql
