+++ common/configuration/scenario/financial_transactions/financial.gql
