+++ common/configuration/scenario/financial_transactions/queries_read.yml
