+++ common/configuration/scenario/road_network/queries_write.yml
