+++ common/configuration/scenario/social_network/queries_read.yml
