+++ common/configuration/scenario/social_network/social_network.gql
