+++ common/configuration/scenario/social_network/social_network_config_read.yml
