+++ profiler/test/ProfilerBootupTestIntegration.java
@@ -18,7 +18,7 @@
 
 package grakn.benchmark.profiler;
 
-import grakn.benchmark.common.configuration.BenchmarkArguments;
+import grakn.benchmark.common.configuration.parse.BenchmarkArguments;
 import grakn.benchmark.common.exception.BootupException;
 import grakn.core.client.GraknClient;
 import grakn.core.concept.answer.ConceptMap;
